refactor: Remove redundant foh and manager contexts in favor of an app context.
