+++ app/routes/_foh+/_layout/header.tsx
@@ -12,7 +12,7 @@ import { ThemeMenu } from '#components/theme-menu';
 import { Button, Divider, Icon, NavLink } from '#components/ui';
 import { UserMenu } from '#components/user-menu';
 import { useIsAuthenticated } from '#utils/auth/user';
-import { useChampionship } from '#utils/foh/championship.context';
+import { useChampionship } from '#utils/foh/use-championship';
 import { usePageTitle } from '#utils/foh/use-page-title';
 
 import { ChampionshipSelect } from './championship-select';
