+++ app/routes/manager+/_layout/_layout.tsx
@@ -1,8 +1,8 @@
 import type { LoaderFunctionArgs } from '@remix-run/node';
-import { Outlet, json } from '@remix-run/react';
+import { Outlet, json, useLoaderData } from '@remix-run/react';
+import { ChampionshipProvider } from '#utils/app/championship.context.js';
 import { requireAdmin } from '#utils/auth/auth.server';
 import { db } from '#utils/db.server';
-import { ChampionshipProvider } from '#utils/manager/championship.context';
 import { ManagerHeader } from './manager-header';
 import { ManagerNav } from './manager-nav';
 
@@ -15,8 +15,9 @@ export async function loader({ request }: LoaderFunctionArgs) {
 }
 
 export default function ManagerLayout() {
+  const { championships } = useLoaderData<typeof loader>();
   return (
-    <ChampionshipProvider>
+    <ChampionshipProvider championships={championships}>
       <div className="grid h-dvh grid-cols-[auto_1fr]">
         <ManagerNav className="w-0 overflow-y-auto border-default border-r bg-app-subtle shadow-medium md:flex md:w-52" />
         <div className="grid grid-rows-[56px_1fr] overflow-y-auto px-2 md:px-4">
